Change hashing algorithm to include interfaces
(cherry picked from commit f683cf8)
